+++ tests/utils/test_entities.py
@@ -229,6 +229,9 @@ orchestrator_1 = Orchestrator(
                 "job_id": "111",
                 "step_id": "111",
             },
+            "client_file": "tests/utils/clients/client.csv",
+            "client_count_file": "tests/utils/clients/client_counts.csv",
+            "memory_file": "tests/utils/memory/memory.csv",
         },
         {
             "name": "shard 2",
@@ -242,6 +245,9 @@ orchestrator_1 = Orchestrator(
                 "job_id": "111",
                 "step_id": "111",
             },
+            "client_file": "tests/utils/clients/client.csv",
+            "client_count_file": "tests/utils/clients/client_counts.csv",
+            "memory_file": "tests/utils/memory/memory.csv",
         },
     ],
 )
@@ -265,6 +271,9 @@ orchestrator_2 = Orchestrator(
                 "job_id": "111",
                 "step_id": "111",
             },
+            "client_file": "tests/utils/clients/client.csv",
+            "client_count_file": "tests/utils/clients/client_counts.csv",
+            "memory_file": "tests/utils/memory/memory.csv",
         },
         {
             "name": "orc 2 shard 2",
@@ -278,6 +287,9 @@ orchestrator_2 = Orchestrator(
                 "job_id": "111",
                 "step_id": "111",
             },
+            "client_file": "tests/utils/clients/client.csv",
+            "client_count_file": "tests/utils/clients/client_counts.csv",
+            "memory_file": "tests/utils/memory/memory_2.csv",
         },
     ],
 )
@@ -300,6 +312,9 @@ orchestrator_3 = Orchestrator(
                 "job_id": "111",
                 "step_id": "111",
             },
+            client_file="tests/utils/clients/client.csv",
+            client_count_file="tests/utils/clients/client_counts.csv",
+            memory_file="tests/utils/memory/memory.csv",
         )
     ],
 )
@@ -322,6 +337,9 @@ orchestrator_4 = Orchestrator(
                 "job_id": "111",
                 "step_id": "111",
             },
+            "client_file": "tests/utils/clients/client.csv",
+            "client_count_file": "tests/utils/clients/client_counts.csv",
+            "memory_file": "tests/utils/memory/memory.csv",
         }
     ],
 )
@@ -344,6 +362,9 @@ no_shards_started = Orchestrator(
                 "job_id": "111",
                 "step_id": "111",
             },
+            "client_file": "tests/utils/clients/client.csv",
+            "client_count_file": "tests/utils/clients/client_counts.csv",
+            "memory_file": "tests/utils/memory/memory.csv",
         },
         {
             "name": "orc 2 shard 2",
@@ -357,6 +378,9 @@ no_shards_started = Orchestrator(
                 "job_id": "111",
                 "step_id": "111",
             },
+            "client_file": "tests/utils/clients/client.csv",
+            "client_count_file": "tests/utils/clients/client_counts.csv",
+            "memory_file": "tests/utils/memory/memory.csv",
         },
     ],
 )
@@ -379,6 +403,9 @@ mismatched_port_orchestrator = Orchestrator(
                 "job_id": "111",
                 "step_id": "111",
             },
+            "client_file": "tests/utils/clients/client.csv",
+            "client_count_file": "tests/utils/clients/client_counts.csv",
+            "memory_file": "tests/utils/memory/memory.csv",
         },
         {
             "name": "shard 2",
@@ -392,6 +419,9 @@ mismatched_port_orchestrator = Orchestrator(
                 "job_id": "111",
                 "step_id": "111",
             },
+            "client_file": "tests/utils/clients/client.csv",
+            "client_count_file": "tests/utils/clients/client_counts.csv",
+            "memory_file": "tests/utils/memory/memory.csv",
         },
     ],
 )
@@ -416,6 +446,9 @@ pending_shard = Shard(
         "job_id": "111",
         "step_id": "111",
     },
+    client_file="tests/utils/clients/client.csv",
+    client_count_file="tests/utils/clients/client_counts.csv",
+    memory_file="tests/utils/memory/memory.csv",
 )
 
 no_return_code_shard = Shard(
@@ -430,6 +463,43 @@ no_return_code_shard = Shard(
         "job_id": "111",
         "step_id": "111",
     },
+    client_file="tests/utils/clients/client.csv",
+    client_count_file="tests/utils/clients/client_counts.csv",
+    memory_file="tests/utils/memory/memory.csv",
+)
+
+telemetry_off_shard = Shard(
+    name="shard 2",
+    hostname="shard2_host",
+    port="11111",
+    out_file="tests/utils/log_files/model_0.out",
+    err_file="tests/utils/log_files/model_0.err",
+    conf_file="/path/to/conf_file",
+    telemetry_metadata={
+        "status_dir": "tests/utils/status_files/model_1",
+        "job_id": "111",
+        "step_id": "111",
+    },
+    client_file="",
+    client_count_file="",
+    memory_file="",
+)
+
+telemetry_files_not_found = Shard(
+    name="shard 2",
+    hostname="shard2_host",
+    port="11111",
+    out_file="tests/utils/log_files/model_0.out",
+    err_file="tests/utils/log_files/model_0.err",
+    conf_file="/path/to/conf_file",
+    telemetry_metadata={
+        "status_dir": "tests/utils/status_files/model_1",
+        "job_id": "111",
+        "step_id": "111",
+    },
+    client_file="not_real",
+    client_count_file="bad_file",
+    memory_file="worse_file",
 )
 
 malformed_status_dir_shard = Shard(
@@ -444,6 +514,9 @@ malformed_status_dir_shard = Shard(
         "job_id": "111",
         "step_id": "111",
     },
+    client_file="tests/utils/clients/client.csv",
+    client_count_file="tests/utils/clients/client_counts.csv",
+    memory_file="tests/utils/memory/memory.csv",
 )
 
 JSONDecoderError_status_shard = Shard(
@@ -458,6 +531,9 @@ JSONDecoderError_status_shard = Shard(
         "job_id": "111",
         "step_id": "111",
     },
+    client_file="tests/utils/clients/client.csv",
+    client_count_file="tests/utils/clients/client_counts.csv",
+    memory_file="tests/utils/memory/memory.csv",
 )
 
 ensemble_1 = Ensemble(