+++ doc/changelog.rst
@@ -6,10 +6,14 @@ Development branch
 
 Description
 
+- Add Database Telemetry page. (SmartDashboard-PR38_)
 - Add Github Actions workflow that checks if changelog is edited
   on pull requests into develop. (SmartDashboard-PR47_)
+- Add manifest file tracking. (SmartDashboard-PR46_)
 
+.. _SmartDashboard-PR38: https://github.com/CrayLabs/SmartDashboard/pull/38
 .. _SmartDashboard-PR47: https://github.com/CrayLabs/SmartDashboard/pull/47
+.. _SmartDashboard-PR46: https://github.com/CrayLabs/SmartDashboard/pull/46
 
 
 0.0.3